% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Prepare species occurrence data from GBIF for climate casting}
\usage{
data_prep(
  gbif_data,
  basis_of_record = NULL,
  coord_unc = NULL,
  identification_verification_status = NULL
)
}
\arguments{
\item{gbif_data}{A data frame holding occurrence records downloaded from GBIF.}

\item{basis_of_record}{Optional character indicating the basisOfRecord types to be included in the data.
If NULL, the default, occurrences with the following basisOfRecord will be kept:  "OBSERVATION", "HUMAN_OBSERVATION",
"MATERIAL_SAMPLE", "LITERATURE", "PRESERVED_SPECIMEN", "UNKNOWN", and "MACHINE_OBSERVATION".}

\item{coord_unc}{Optional numeric indicating the maximal coordinate uncertainty (m) an
observation can have to be included in the data.
If NULL, the default, all occurrences will be kept regardless of their coordinate uncertainty.}

\item{identification_verification_status}{Optional character or a character vector indicating the identificationVerificationStatus of occurrence records that will be kept.
If NULL, the default, all occurrences will be kept except those with the following identificationVerificationStatus: "unverified", "unvalidated", "not validated", "under validation", "not able to validate", "control could not be conclusive due to insufficient knowledge",  "Control could not be conclusive due to insufficient knowledge", "1","uncertain", "unconfirmed", "Douteux", "Invalide", "Non réalisable", "verification needed" , "Probable", "unconfirmed - not reviewed", "validation requested".}
}
\value{
An sf data frame holding occurrence records that are ready to be used for climate casting.
}
\description{
Prepare species occurrence data from GBIF for climate casting
}
\examples{
# Create a dataframe with occurrence data
 example_df <- rgbif::occ_download_get("0001221-210914110416597",overwrite = TRUE) \%>\%
               rgbif::occ_download_import()
 df<-data_prep(example_df)

}
