% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stitch.R
\name{data_stitch}
\alias{data_stitch}
\title{Combine occurrence data from different species into a single species complex and prepare this dataset for climate casting.}
\usage{
data_stitch(x, to_stitch, taxa_complex_name, taxa_complex_key = NULL)
}
\arguments{
\item{x}{An sf dataframe, resulting from either get_gbif_data(), get_zip_data(), or get_downloadkey_data().}

\item{to_stitch}{A numeric vector of two or more taxon keys, present in x under "acceptedTaxonKey", that should be combined into a species complex.}

\item{taxa_complex_name}{A character value, specifying the name that will be assigned to the species complex.}

\item{taxa_complex_key}{An optional numeric value, specifying the acceptedTaxonKey that will be assigned to the species complex. If not provided,the taxon key of one of the species in the complex is used.}
}
\value{
A sf data frame, holding occurrence records of a species complex, ready to be used for climate casting.
}
\description{
Combine occurrence data from different species into a single species complex and prepare this dataset for climate casting.
}
\examples{
 \dontrun{
taxon_key <- c(2865504,2891932,2891930)
df<-get_gbif_data(taxon_key,
                  path = tempdir()
                  )
df_stitched<- data_stitch(df,
                          to_stitch = c(2891932,2891930),
                          taxa_complex_name= "Casuarina cunninghamiana/equisetifolia"
                          )
          }
}
