% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_thin.R
\name{data_thin}
\alias{data_thin}
\title{Data thining function}
\usage{
data_thin(df, cutoff = 1000, fun = "mean")
}
\arguments{
\item{cutoff}{A numeric value defining the distance cutoff for grouping points}

\item{fun}{A function to apply to the grouped points. Default is 'mean'}

\item{data}{A sf object resulting from previous steps in the ClimateCastR - flow}
}
\value{
A sf object with the same columns as the input data, but with fewer rows
}
\description{
A function to reduce the number of data points in a data frame by proximity.
The function groups points that are close to each other. The group is then
represented by the mean of the points in the group. Closeness is defined by a
spatial cutoff distance.
}
\details{
The cutoff is set to 1000 meters by default. This means that points that are
within 1000 meters of each other will be grouped together. The cutoff can be
changed by setting the cutoff argument to a different value.

The fun argument can be any function that takes a numeric vector as input and
returns a single numeric value. The default is the mean function. Other options
are median, min, max & sum.
}
\examples{
\dontrun{
# get data using get_data functions or as a result of the data_prep function.
taxon_key <- c(2865504, 5274858)
df<-get_gbif_data(taxon_key)
}

}
