% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_gbif_data}
\alias{get_gbif_data}
\title{Download species occurrence data from GBIF and prepare them for climate casting}
\usage{
get_gbif_data(
  taxon_key,
  basis_of_record = NULL,
  coord_unc = NULL,
  identification_verification_status = NULL,
  region_shape = NULL
)
}
\arguments{
\item{taxon_key}{A numeric value or vector specifying the GBIF taxonKey(s).}

\item{basis_of_record}{Optional character indicating the basisOfRecord types to be included in the data.
If NULL, the default, occurrences with the following basisOfRecord will be kept:  "OBSERVATION", "HUMAN_OBSERVATION",
"MATERIAL_SAMPLE", "LITERATURE", "PRESERVED_SPECIMEN", "UNKNOWN", and "MACHINE_OBSERVATION".}

\item{coord_unc}{Optional numeric indicating the maximal coordinate uncertainty (m) an
observation can have to be included in the data.
If NULL, the default, all occurrences will be kept regardless of their coordinate uncertainty.}

\item{identification_verification_status}{Optional character or a character vector indicating the identificationVerificationStatus of occurrence records that will be kept.
If NULL, the default, all occurrences will be kept except those with the following identificationVerificationStatus: "unverified", "unvalidated", "not validated", "under validation", "not able to validate", "control could not be conclusive due to insufficient knowledge",  "Control could not be conclusive due to insufficient knowledge",
"1","uncertain", "unconfirmed", "Douteux", "Invalide", "Non réalisable", "verification needed" , "Probable", "unconfirmed - not reviewed", "validation requested".}

\item{region_shape}{Optional character specifying a path to a shape file indicating the region from which occurrence records should be downloaded.}
}
\value{
An sf data frame holding occurrence records that are ready to be used for climate casting.
}
\description{
This function downloads species occurrence records from GBIF, filters them,
and converts them into an sf data frame which can then be used for climate casting.
}
\examples{
 \dontrun{
#provide GBIF taxon_key(s)
taxon_key <- c(2865504, 5274858)
get_gbif_data(taxon_key)
get_gbif_data(taxon_key,
              coord_unc = 100,
              basis_of_record = "HUMAN_OBSERVATION",
              identification_verification_status= c("Probable", "valid","confident","")
              )

#provide a taxon key and a path to a shapefile
taxon_key<-2427091
url <- "https://zenodo.org/records/3386224"
zipfile <- tempfile(fileext = ".zip")
download.file(url, zipfile, mode = "wb")
unzip(zipfile, exdir = tempdir())
shapefile_path<-file.path(tempdir(), "flanders.shp")

get_gbif_data(taxon_key,
             region_shape = shapefile_path
             )
}

}
\author{
Soria Delva, Sander Devisscher
}
