% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_zip_data}
\alias{get_zip_data}
\title{Import species occurrence data from a zip file and prepare them for climate casting}
\usage{
get_zip_data(
  zip_file,
  basis_of_record = NULL,
  coord_unc = NULL,
  identification_verification_status = NULL
)
}
\arguments{
\item{zip_file}{Character specifying the path to a zip file holding occurrence records from a previous GBIF download.}

\item{basis_of_record}{Optional character indicating the basisOfRecord types to be included in the data.
If NULL, the default, occurrences with the following basisOfRecord will be kept:  "OBSERVATION", "HUMAN_OBSERVATION",
"MATERIAL_SAMPLE", "LITERATURE", "PRESERVED_SPECIMEN", "UNKNOWN", and "MACHINE_OBSERVATION".}

\item{coord_unc}{Optional numeric indicating the maximal coordinate uncertainty (m) an
observation can have to be included in the data.
If NULL, the default, all occurrences will be kept regardless of their coordinate uncertainty.}

\item{identification_verification_status}{Optional character or a character vector indicating the identificationVerificationStatus of occurrence records that will be kept.
If NULL, the default, all occurrences will be kept except those with the following identificationVerificationStatus: "unverified", "unvalidated", "not validated", "under validation", "not able to validate", "control could not be conclusive due to insufficient knowledge",  "Control could not be conclusive due to insufficient knowledge", "1","uncertain", "unconfirmed", "Douteux", "Invalide", "Non réalisable", "verification needed" , "Probable", "unconfirmed - not reviewed", "validation requested".}
}
\value{
An sf data frame holding occurrence records that are ready to be used for climate casting.
}
\description{
This function imports species occurrence data from a zip file, filters them,
and converts them into an sf data frame which can then be used for climate casting.
}
\examples{
#'  \dontrun{
#download zip_file from GBIF
zipdir<-tempdir()
rgbif::occ_download_get("0001221-210914110416597",
                      path = zipdir,
                      overwrite = TRUE)
zip_file <- file.path(zipdir, "0001221-210914110416597.zip")
file.exists(zip_file)
df<-get_zip_data(zip_file)
}

}
\author{
Soria Delva, Sander Devisscher
}
